Split rawbuffer tracking from ptr info
Add an optimizeopt/rawbuffer.rs module for the RPython rawbuffer.py port, with RawBuffer owning the four parallel offset/length/descr/value lists and preserving the sorted non-overlapping read/write semantics.

Keep RawBufferPtrInfo parity in info.rs by making VirtualRawBufferInfo own a RawBuffer field, matching PyPy's RawBufferPtrInfo.buffer shape instead of merging the raw buffer lists directly into the pointer-info metadata.

Update raw malloc virtualization, virtual-state walking, forcing, and tests to go through the nested buffer while preserving existing public optimizer behavior and the documented fixed-width integer overflow adaptations.

Tests: cargo check --features dynasm; cargo test -p majit-metainterp --features dynasm rawbuffer; cargo test --features dynasm
